完善页面逻辑

+++ src/pages/user/advance-order/advanceOrder.js
@@ -1,8 +1,8 @@
-import Taro from '@tarojs/taro'
-import {Image, Text, View} from '@tarojs/components'
-import {AtTabs, AtTabsPane,} from "taro-ui"
+import Taro,{eventCenter} from '@tarojs/taro'
+import {Button, Image, Text, View} from '@tarojs/components'
+import {AtModal, AtModalAction, AtTabs, AtTabsPane,} from "taro-ui"
 import './advanceOrder.scss'
-import {getMyAppointListApi} from "../../../services/user";
+import {getMyAppointListApi, queryAppointRecord} from "../../../services/user";
 import moment from "moment";
 import React, {Component} from 'react'
 import Api from '../../../config/api';
@@ -11,19 +11,24 @@ import Forward from '@assets/home/forward.svg'
 import Config from "../../../../project.config.json";
 import _Empty from "@assets/empty.png";
 
+let i = 1;
+
 export class AdvanceOrder extends Component {
   state = {
     current: 0,
     list: [],
     state: '',
     page: 1,
+    item:{},
     limit: 10,
     totalPage: 1,
+    visible:false,
     userId: '',
     tabList: [{title: '全部', id: 0}, {title: '预约中', id: 1}, {title: '已预约', id: 2}, {title: '已完成', id: 3}]
   }
 
   componentDidMount() {
+
     this._initData();
   }
 
@@ -68,7 +73,6 @@ export class AdvanceOrder extends Component {
     })
 
   }
-
   onReachBottom() {
     if (this.state.totalPage > this.state.page) {
       this.setState({
@@ -86,7 +90,6 @@ export class AdvanceOrder extends Component {
       return false;
     }
   }
-
   handleClick = (value) => {
     let state = '';
     switch (value) {
@@ -106,28 +109,16 @@ export class AdvanceOrder extends Component {
         state = 3;
         break;
     }
+    console.log(333,value);
     this.setState({current: value, state, page: 1, list: []}, () => {
       this._getList();
     })
   }
   goToPage = (item) => {
-    console.log(333,item);
+    console.log(333, item);
     if (item.state == 1) {
       Taro.navigateTo({
         url: `/pages/user/order-success/orderAppointSuccess?item=${JSON.stringify(item)}`,
-        events: {
-          // 为指定事件添加一个监听器，获取被打开页面传送到当前页面的数据
-          acceptDataFromOpenedPage: function (data) {
-            console.log(data)
-          },
-          someEvent: function (data) {
-            console.log(data)
-          }
-        },
-        success: function (res) {
-          // 通过eventChannel向被打开页面传送数据
-          res.eventChannel.emit('acceptDataFromOpenerPage', {data: 'test'})
-        }
       })
     }
   }
@@ -151,42 +142,63 @@ export class AdvanceOrder extends Component {
     }
   }
   _orderOp = async (item) => {
-    item.state == 1 && Taro.request({
-      url: Api.cancelOrder + `?id=${item.id}`, //仅为示例，并非真实的接口地址
-      data: {},
-      method: 'POST',
-      header: {
-        'content-type': 'application/json' // 默认值
-      },
-      success: (res) => {
-        console.log(333, res);
-        const _res = res.data;
-        _res.code == 200 && this.setState({page: 1, list: []}, () => {
-          this._getList();
+    if(item.state ==1){
+      this.setState({visible:true,item});
+    }else {
+        const res = await queryAppointRecord({
+          id: item.id,
         })
-      }
-    })
+        if(res.code == 200){
+          const {userType}=res.data;
+          Taro.eventCenter.trigger('event', res.data)
+            Taro.redirectTo({url:`/pages/home/organization/organization?userType=${userType}`});
+        }
+
+    }
 
   }
   _deleteOrder = (item) => {
-    Taro.request({
-      url: Api.deleteOrder + `?id=${item.id}`, //仅为示例，并非真实的接口地址
-      data: {},
-      method: 'DELETE',
-      header: {
-        'content-type': 'application/json' // 默认值
-      },
-      success: (res) => {
-        const _res = res.data;
-        _res.code == 200 && this.setState({page: 1, list: []}, () => {
-          this._getList();
-        })
-      }
-    })
+    this.setState({visible:true,item})
+  }
+  _enter=()=>{
+    const {item}=this.state;
+    if(item.state==3) {
+      Taro.request({
+        url: Api.deleteOrder + `?id=${item.id}`, //仅为示例，并非真实的接口地址
+        data: {},
+        method: 'DELETE',
+        header: {
+          'content-type': 'application/json' // 默认值
+        },
+        success: (res) => {
+          const _res = res.data;
+          console.log(222,_res);
+          _res.code == 200 && this.setState({visible: false, page: 1, list: []}, () => {
+            this._getList();
+          })
+        }
+      })
+    }else if(item.state==1){
+      Taro.request({
+        url: Api.cancelOrder + `?id=${item.id}`, //仅为示例，并非真实的接口地址
+        data: {},
+        method: 'POST',
+        header: {
+          'content-type': 'application/json' // 默认值
+        },
+        success: (res) => {
+          console.log(333, res);
+          const _res = res.data;
+          _res.code == 200 && this.setState({visible:false,page: 1, list: []}, () => {
+            this._getList();
+          })
+        }
+      })
+    }
   }
 
   render() {
-    const {current, tabList, list} = this.state;
+    const {current, tabList, list,visible,item:tipItem} = this.state;
     console.log(333, list);
     return (
       <View className='container'>
@@ -194,41 +206,53 @@ export class AdvanceOrder extends Component {
           <AtTabs current={current} tabList={tabList} onClick={this.handleClick}>
             {tabList.map((item, index) => {
               return (
-                <AtTabsPane  className='at-tab' key={item.id + ""} current={current} index={index}>
-                  {list.length!==0?list.map((_item, index) => {
-                      let date = moment(_item.date).format('YYYY-MM-DD');
-                      let week = this._getWeek(_item.date);
-                      return (
-                        <View className='wrap' key={_item.id + " "} >
-                          <View className='main'>
-                            <View className='listItem'>
-                              <View className='listItem_left'>
-                                <Text className='listItem_left_appoint'>预约人:{_item.name}</Text>
-                                <Text className='listItem_left_date'>{date} {week}</Text>
-                              </View>
-                              <View className='listItem_right' onClick={(item) => this.goToPage(_item)}>
-                                <Text
-                                  className='listItem_right_status'>{_item.state == 0 ? '预约中' : _item.state == 1 ? '已预约' : _item.state == 2 ? '已完成' : '已取消'}</Text>
-                                <Image src={Forward} className='listItem_right_arrow'/>
-                              </View>
+                <AtTabsPane className='at-tab' key={item.id + ""} current={current} index={index}>
+                  {list.length !== 0 ? list.map((_item, index) => {
+                    let date = moment(_item.date).format('YYYY-MM-DD');
+                    let week = this._getWeek(_item.date);
+                    console.log(333, _item);
+                    return (
+                      <View className='wrap' key={_item.id + " "}>
+                        <View className='main'>
+                          <View className='listItem'>
+                            <View className='listItem_left'>
+                              <Text className='listItem_left_appoint'>预约人:{_item.name}</Text>
+                              <Text className='listItem_left_date'>{date} {week} {_item.timeType==0?'上午':_item.timeType==1?'下午':'全天'}</Text>
                             </View>
-                            <View className='footer' >
-                              <View className='op_btn_1' onClick={() => this._orderOp(_item)}>
-                                <Text>{_item.state == 1 ? '取消预约' : '再次预约'}</Text>
-                              </View>
-                              {_item.state == 3 && <View className='op_btn_2' onClick={() => this._deleteOrder(_item)}>
-                                <Text>删除</Text>
-                              </View>}
+                            <View className='listItem_right' onClick={(item) => this.goToPage(_item)}>
+                              <Text
+                                className='listItem_right_status' style={_item.state==0?'color:red':_item.state==1?'color:green':_item.state==2?'#333':'#999'}>{_item.state == 0 ? '预约中' : _item.state == 1 ? '已预约' : _item.state == 2 ? '已完成' : '已取消'}</Text>
+                              <Image src={Forward} className='listItem_right_arrow'/>
                             </View>
                           </View>
+                          <View className='footer'>
+                            <View className='op_btn_1' onClick={() => this._orderOp(_item)}>
+                              <Text>{_item.state == 1 ? '取消预约' : '再次预约'}</Text>
+                            </View>
+                            {_item.state == 3 && <View className='op_btn_2' onClick={() => this._deleteOrder(_item)}>
+                              <Text>删除</Text>
+                            </View>}
+                          </View>
                         </View>
-                      )
-                    }):<Empty/>}
+                      </View>
+                    )
+                  }) : <Empty/>}
                 </AtTabsPane>
               )
             })}
           </AtTabs>
         </View>
+        <AtModal
+          isOpened={visible}
+        >
+          <View className='modal-view'>
+            <Text className='modal-text'>确定{tipItem.state==3?'删除':'取消'}该条预约信息吗？</Text>
+          </View>
+          <AtModalAction>
+            <Button className={'btn'} onClick={()=>this.setState({visible:false})}>取消</Button>
+            <Button onClick={this._enter}>确定</Button>
+          </AtModalAction>
+        </AtModal>
       </View>
     )
   }
